Replace convert_list_to_string() with "\n".join() (#5)
"\n".join() also concatenates list elements with a line break
in between each, but omits the last line break. This functionality
is fully equivalent within embed fields, as line breaks at the end of
fields do not add new lines (we would need a Zero-width character if
we wanted to accomplish this).