Unify queue methods (#8)
* Combine remove_from_queue() and leave_queue()

The methods are functionally equivalent, apart from their signatures.
The ctx is not needed though, passing the Member is enough. Since the
queue is a list of Members, take the Member as an arg.

* Let join_queue() take a Member instead of ctx

Only ctx.user is needed, so pass this instead.

* Remove unused imports

* Remove rogue comment

* Rename queue methods to add() and remove()

Now that they are queue methods, specifying the object of the
operation as "queue" is no longer needed. From a queue's perspectives,
we are adding and removing users.

---------

Co-authored-by: Kenyon Allan <48738848+kenyon-allan@users.noreply.github.com>